+++ src/main/__tests__/helpers/oauthHelpers.ts
@@ -1,11 +1,12 @@
-import { ResponseType } from '../../authOptions'
 import { toQueryString } from '../../../utils/queryString'
 import { pick } from '../../../utils/utils'
+import { ResponseType } from '../../authOptions'
 import { AuthParameters } from '../../authParameters'
+import { WithPkceParams } from '../../pkceService'
 
 export const mockPkceValues = {
   code_challenge: 'KBHQASQDHCtWjDGVGQaPjsK8c8SlrH2yfm3nQh75o14',
-  code_challenge_method: 'S256',
+  code_challenge_method: 'S256'
 }
 
 export const defaultScope = 'openid profile email phone'
@@ -37,23 +38,23 @@ export const confidential = { isPublic: false }
 
 export const webMessage = { responseMode: 'web_message', prompt: 'none' }
 
-export const pageDisplay = { display: 'page' }
-export const popupDisplay = { display: 'popup' }
+export const pageDisplay: Pick<AuthParameters, 'display'> = { display: 'page' }
+export const popupDisplay: Pick<AuthParameters, 'display'> = { display: 'popup' }
 
-type ExpectedPkceResult<IsPublic, ResponseType> =
-  IsPublic extends true ?
-    ResponseType extends 'code' ?
-        typeof mockPkceValues :
-        Record<string, never> :
-    Record<string, never>
+type ExpectedPkceResult<IsPublic, ResponseType> = IsPublic extends true
+  ? ResponseType extends 'code'
+    ? typeof mockPkceValues
+    : Record<string, never>
+  : Record<string, never>
 
-function expectedPkce(isPublic: boolean, responseType: ResponseType): ExpectedPkceResult<typeof isPublic, typeof responseType> {
-  return (isPublic && responseType === 'code')
-    ? mockPkceValues
-    : {}
+function expectedPkce(
+  isPublic: boolean,
+  responseType: ResponseType
+): ExpectedPkceResult<typeof isPublic, typeof responseType> {
+  return isPublic && responseType === 'code' ? mockPkceValues : {}
 }
 
-export function getExpectedQueryString(params: ClientType & AuthParameters & Partial<TknType>): string {
+export function getExpectedQueryString(params: ClientType & WithPkceParams<AuthParameters> & Partial<TknType>): string {
   const responseType = params['responseType']
   const isPublic = params['isPublic']
   const pkceValues = expectedPkce(isPublic, responseType)
@@ -66,11 +67,13 @@ export function getExpectedQueryString(params: ClientType & AuthParameters & Par
     ...pick(params, 'state'),
     ...pick(params, 'nonce'),
     ...pick(params, 'providerScope'),
+    ...pick(params, 'codeChallenge'),
+    ...pick(params, 'codeChallengeMethod'),
     ...scope,
     ...pick(params, 'display'),
     ...pick(params, 'responseMode'),
     ...pick(params, 'prompt'),
     ...pkceValues,
-    ...pick(params, 'tkn'),
+    ...pick(params, 'tkn')
   })
 }