+++ src/main/models.ts
@@ -410,6 +410,11 @@ export type SessionDevice = {
 export type SessionDeviceListResponse = {
   sessionDevices: SessionDevice[]
 }
+export type LoginTypeAllowed = {
+  email: boolean
+  phoneNumber: boolean
+  customIdentifier: boolean
+}
 /**
  * This type represents the settings of a ReachFive account's stored in the backend.
  */
@@ -432,6 +437,7 @@ export type RemoteSettings = {
   mfaEmailEnabled: boolean,
   rbaEnabled: boolean,
   isImplicitFlowForbidden: boolean
+  loginTypeAllowed: LoginTypeAllowed
 }
 
 export type ErrorResponse = {