+++ src/main/pkceService.ts
@@ -5,11 +5,13 @@ import { randomBase64String } from '../utils/random'
 
 export type PkceParams = { codeChallenge: string; codeChallengeMethod: string }
 
+export type WithPkceParams<T> = T & Partial<PkceParams>
+
 export function computePkceParams(): Promise<PkceParams> {
   const codeVerifier = randomBase64String()
 
   localStorage.setItem('verifier_key', codeVerifier)
-  return computeCodeChallenge(codeVerifier).then(challenge => {
+  return computeCodeChallenge(codeVerifier).then((challenge) => {
     return {
       codeChallenge: challenge,
       codeChallengeMethod: 'S256'
@@ -19,8 +21,8 @@ export function computePkceParams(): Promise<PkceParams> {
 
 function computeCodeChallenge(verifier: string): Promise<string> {
   const binaryChallenge = Buffer.from(verifier, 'utf-8')
-    return new Promise(resolve => {
-    window.crypto.subtle.digest('SHA-256', binaryChallenge).then(hash => {
+  return new Promise((resolve) => {
+    window.crypto.subtle.digest('SHA-256', binaryChallenge).then((hash) => {
       return resolve(encodeToBase64(hash))
     })
   })